modified files for customer modularization
