+++ project/src/templates/rewards/view_game_board.j2
@@ -208,18 +208,5 @@
 	</div>
 	
 	<script src="{{ url_for('static', filename='js/view_board.js') }}"></script>
-	<script src="{{ url_for('static', filename='js/reset_board.js') }}"></script>
-	<script>
-		function myFunction() {
-			if (
-			confirm(
-				// get user confirmation
-				'Are you sure you want to leave this page? All unsaved changes will be lost.'
-			)
-		) {
-			history.back(); // redirect user
-		}
-	</script>
-
 	
 {% endblock body %}
